Add abbility to set value for title tab from Admin
