+++ etc/config.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
+    <default>
+        <animationTitleTab>
+            <general>
+                <enable>0</enable>
+                <getValueFromAdmin>0</getValueFromAdmin>
+            </general>
+        </animationTitleTab>
+    </default>
+</config>
