+++ view/frontend/layout/catalog_category_view.xml
@@ -2,6 +2,15 @@
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <head>
-        <script src="Zvirko_AnimateTitle::js/animateTitle.js"/>
+        <script src="Zvirko_AnimateTitle::js/animationTitleTab.js"/>
     </head>
+    <body>
+        <referenceContainer name="content">
+            <block template="Zvirko_AnimateTitle::config.phtml">
+                <arguments>
+                    <argument name="view_model" xsi:type="object">Zvirko\AnimateTitle\ViewModel\Config</argument>
+                </arguments>
+            </block>
+        </referenceContainer>
+    </body>
 </page>
